LayoutTests/inspector/dom/setEventListenerDisabled-expected.txt

Testing DOMAgent.setEventListenerDisabled.


== Running test suite: DOM.setEventListenerDisabled
-- Running test case: DOM.setEventListenerDisabled.DisabledClickEvent
Click event listener is enabled.
Disabling event listener...
PASS: Click event listener did not fire.
<body> clicked.
Click event listener is disabled.

-- Running test case: DOM.setEventListenerDisabled.ReenabledClickEvent
Click event listener is disabled.
Enabling event listener...
PASS: Click event listener fired.
<body> clicked.
Click event listener is enabled.

-- Running test case: DOM.setEventListenerDisabled.Invalid
PASS: Should produce an error.
Error: No event listener for given identifier.
